Split test-shell.ts to separate concerns
